Added tests around the virgil message conversion
